refactor: 악곡 상세 컴포넌트 구조 분리
